lambdas/aspect-calculator/requirements.txt
boto3==1.35.0
psycopg2-binary==2.9.9
numpy==1.26.4
GDAL==3.8.5
matplotlib==3.9.2
